Fixed Bug #98450 - inconsistencies between frontend DB and incoming files.
Introduced a new qc_status "File not found".
The "File not found" status is assigned to deliveries which do not have
a ZIP file in the expected filesystem path.
Delivery entries  with "File not found" status can be deleted by the user.

Also added handling of situation when a file with the same name as an uploaded
file exists in the file system, but is not registered in the DB. In this case
the original file is replaced by the uploaded file.