Update status of all running jobs in the db on page reload.
Before reading contents of the deliveries table from the database,
check status of all jobs that are marked in the database as "running".

The status of running jobs is also done on the client but it has only
been done for jobs on the first visible page. Updating the status server-side
ensures that jobs on invisible pages are also up-to-date in the database.