raster.mmu: number of reported error and exception patches is limited to 1000000 threshold.
If more than one million patches < MMU are found, further detection of patches with area < MMU is stopped and only the first 1,000,000 error patch locations are added to the attachment.
This is to prevent excessive memory and disk usage in the case of a very large number of error patches.