Raster.gap check supports AOI in vector format.
The AOI vector file must be located in BOUNDARY_DIR/raster/<mask_ident.gpkg> where mask_ident.gpkg is the value of the 'mask' parameter.
The du_column_name parameter also needs to be specified to tell the check which attribute it should look for in the AOI vector file.
The AOI vector file is first filtered by aoi_code (extracted in raster.naming check) and then the filter polygons are rasterized.
The generated raster mask is then used in gap detection.