raster.origin check - added a customizable grid_size parameter.
The grid_size parameter specifies the size of the grid (in units of the raster crs) that the raster origin should be exactly aligned to.
The default value of grid_size is 1000 metres (LEAC 1km grid).
If the coordinates of the raster origin  should be integers, then grid_size parameter should be set to 1 metre.