Ensure that bounding box of mask and checked raster are correctly aligned.
When the raster.gap mask is generated on-the-fly and the raster coordinates are not a multiple of the pixel size, it is ensured that the origin of the new mask is aligned to the origin of the checked raster.
The origin of the mask is placed to ensure that:
abs(mask_ulx - raster_ulx) is divisible by pixel_size with no remainder.
abs(mask_uly - raster_uly) is divisible by pixel_size with no remainder.