Introduced JOB_LOST and JOB_TIMEOUT statuses.
JOB_LOST ("worker lost") is shown when a worker gets shut down in the middle of running a job.
JOB_TIMEOUT ("worker timeout") is shown when a worker is too slow to respond to a request about status of a running job.
The new statuses have been introduced to troubleshoot unspecified "error" statuses appearing ocasionally on eea qctool instance.